feat: add malloc configuration through the env
The library is initialized with a cool C feature I did not know about.
The constructor and destructor attributes for library functions. If a
function is created with this attribute it will be called when the
library is loaded which means we can initialize a global state for the
library if needed. Then we can clean it with the destructor function.
This is great!